Bedtime_Story/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
edge-tts>=7.0.0 